@@ -5,8 +5,8 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "test:unit": "vue-cli-service test:unit"
+    "test:unit": "vue-cli-service test:unit",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
     "@vue/composition-api": "^0.5.0",
@@ -19,10 +19,10 @@
     "jsonp": "^0.2.1",
     "lyric-parser": "^1.0.1",
     "register-service-worker": "^1.7.1",
-    "vue": "^2.6.11",
+    "vue": "^3.0.0-beta.14",
     "vue-lazyload": "^1.3.3",
-    "vue-router": "^3.1.6",
-    "vuex": "^3.1.3"
+    "vue-router": "^4.0.0-alpha.12",
+    "vuex": "^4.0.0-beta.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^4.3.0",
@@ -32,17 +32,18 @@
     "@vue/cli-plugin-unit-jest": "^4.3.0",
     "@vue/cli-plugin-vuex": "^4.3.0",
     "@vue/cli-service": "^4.3.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.14",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "1.0.0-beta.32",
+    "@vue/test-utils": "^2.0.0-alpha.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.8.0",
     "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "lint-staged": "^10.1.2",
-    "prettier": "^1.18.2",
+    "prettier": "1.19.1",
     "sass": "^1.26.3",
     "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.11"
+    "vue-cli-plugin-vue-next": "~0.1.2"
   },
   "gitHooks": {
     "pre-commit": "lint-staged"
