@@ -1,95 +1,92 @@
-import {
-  onMounted,
-  watch,
-  computed,
-  onActivated,
-  reactive
-} from '@vue/composition-api'
-import { playMode } from '../utils/config'
-import { shuffle } from '../utils/util'
+import { onMounted, watch, computed, onActivated, reactive, ref } from "vue";
+import { useStore } from "vuex";
+import { playMode } from "../utils/config";
+import { shuffle } from "../utils/util";
 
-export function usePlaylist(vm, handlePlaylist) {
-  const playlist = computed(() => vm.$store.getters.playlist)
+export function usePlaylist(handlePlaylist) {
+  const store = useStore();
+  const playlist = computed(() => tore.getters.playlist);
 
   onMounted(() => {
-    handlePlaylist(playlist)
-  })
+    handlePlaylist(playlist);
+  });
 
   onActivated(() => {
-    handlePlaylist(playlist)
-  })
+    handlePlaylist(playlist);
+  });
 
   watch(
     () => playlist,
-    newVal => handlePlaylist(newVal),
-    { lazy: true }
-  )
+    (newVal) => handlePlaylist(newVal)
+  );
 
-  return {}
+  return {};
 }
 
-export function usePlayer(vm) {
-  const sequenceList = computed(() => vm.$store.getters.sequenceList)
-  const currentSong = computed(() => vm.$store.getters.currentSong)
-  const playlist = computed(() => vm.$store.getters.playlist)
-  const mode = computed(() => vm.$store.getters.mode)
-  const favoriteList = computed(() => vm.$store.getters.favoriteList)
+export function usePlayer() {
+  const store = useStore();
+  const sequenceList = computed(() => store.getters.sequenceList);
+  const currentSong = computed(() => store.getters.currentSong);
+  const playlist = computed(() => store.getters.playlist);
+  const mode = computed(() => store.getters.mode);
+  const favoriteList = computed(() => store.getters.favoriteList);
 
-  const setPlayingState = useMutations(vm, 'SET_PLAYING_STATE')
-  const setCurrentIndex = useMutations(vm, 'SET_CURRENT_INDEX')
-  const setPlayMode = useMutations(vm, 'SET_PLAY_MODE')
-  const setPlaylist = useMutations(vm, 'SET_PLAYLIST')
+  const setPlayingState = (flag) => store.commit("SET_PLAYING_STATE", flag);
+  const setCurrentIndex = (index) => store.commit("SET_CURRENT_INDEX", index);
+  const setPlayMode = (mode) => store.commit("SET_PLAY_MODE", mode);
+  const setPlaylist = (list) => store.commit("SET_PLAYLIST", list);
 
-  const saveFavoriteList = useActions(vm, 'saveFavoriteList')
-  const deleteFavoriteList = useActions(vm, 'deleteFavoriteList')
+  const saveFavoriteList = (song) => store.dispatch("saveFavoriteList", song);
+  const deleteFavoriteList = (song) =>
+    store.dispatch("deleteFavoriteList", song);
 
   const iconMode = computed(() => {
     return mode.value === playMode.sequence
-      ? 'icon-sequence'
+      ? "icon-sequence"
       : mode.value === playMode.loop
-      ? 'icon-loop'
-      : 'icon-random'
-  })
+      ? "icon-loop"
+      : "icon-random";
+  });
 
-  const favoriteIcon = computed(() => getFavoriteIcon(currentSong.value))
+  const favoriteIcon = computed(() => getFavoriteIcon(currentSong.value));
 
   function changeMode() {
-    const mode = (mode.value + 1) % 3
-    setPlayMode(mode)
+    const mode = (mode.value + 1) % 3;
+    setPlayMode(mode);
 
-    let list
+    let list;
     if (mode.value === playMode.random) {
-      list = shuffle(sequenceList.value)
+      list = shuffle(sequenceList.value);
     } else {
-      list = sequenceList.value
+      list = sequenceList.value;
     }
 
-    resetCurrentIndex(list)
-    setPlaylist(list)
+    resetCurrentIndex(list);
+    setPlaylist(list);
   }
 
   function resetCurrentIndex(list) {
     const currentIndex = list.findIndex(
-      item => item.id === currentSong.value.id
-    )
-    setCurrentIndex(currentIndex)
+      (item) => item.id === currentSong.value.id
+    );
+    setCurrentIndex(currentIndex);
   }
 
   function toggleFavorite(song) {
     if (isFavorite(song)) {
-      deleteFavoriteList(song)
+      deleteFavoriteList(song);
     } else {
-      saveFavoriteList(song)
+      saveFavoriteList(song);
     }
   }
 
   function getFavoriteIcon(song) {
-    return isFavorite(song) ? 'icon-favorite' : 'icon-not-favorite'
+    return isFavorite(song) ? "icon-favorite" : "icon-not-favorite";
   }
 
   function isFavorite(song) {
-    const index = favoriteList.value.findIndex(item => item.id === song.id)
-    return index > -1
+    const index = favoriteList.value.findIndex((item) => item.id === song.id);
+    return index > -1;
   }
 
   return {
@@ -108,56 +105,49 @@ export function usePlayer(vm) {
     favoriteIcon,
     changeMode,
     getFavoriteIcon,
-    toggleFavorite
-  }
+    toggleFavorite,
+  };
 }
 
-export function useSearch(vm, refs) {
-  const saveSearchHistory = useActions(vm, 'saveSearchHistory')
-  const deleteSearchHistory = useActions(vm, 'deleteSearchHistory')
+export function useSearch() {
+  const searchBoxRef = ref(null);
 
-  const _state = reactive({
-    query: '',
-    refreshDelay: 120
-  })
+  const saveSearchHistory = (query) =>
+    store.dispatch("saveSearchHistory", query);
+  const deleteSearchHistory = (query) =>
+    store.dispatch("deleteSearchHistory", query);
 
-  const searchHistory = computed(() => vm.$store.getters.searchHistory)
+  const searchState = reactive({
+    query: "",
+    refreshDelay: 120,
+  });
+
+  const searchHistory = computed(() => vm.$store.getters.searchHistory);
 
   function addQuery(query) {
-    refs.searchBoxRef.setQuery(query)
+    searchBoxRef.setQuery(query);
   }
 
   function blurInput() {
-    refs.searchBoxRef.blur()
+    searchBoxRef.blur();
   }
 
   function saveSearch() {
-    saveSearchHistory(_state.query)
+    saveSearchHistory(searchState.query);
   }
 
   function onQueryChange(query) {
-    _state.query = query.trim()
+    searchState.query = query.trim();
   }
 
   return {
-    _state,
+    searchBoxRef,
+    searchState,
     searchHistory,
     addQuery,
     saveSearch,
     onQueryChange,
     deleteSearchHistory,
-    blurInput
-  }
-}
-
-export function useMutations(vm, type) {
-  return payload => {
-    vm.$store.commit(type, payload)
-  }
-}
-
-export function useActions(vm, type) {
-  return payload => {
-    vm.$store.dispatch(type, payload)
-  }
+    blurInput,
+  };
 }
