@@ -1,17 +1,17 @@
 // 基本参数
 export const commonParams = {
   g_tk: 1928093487,
-  inCharset: 'utf-8',
-  outCharset: 'utf-8',
+  inCharset: "utf-8",
+  outCharset: "utf-8",
   notice: 0,
-  format: 'jsonp'
-}
+  format: "jsonp",
+};
 
 // jsonp 参数选项
 export const options = {
-  param: 'jsonpCallback',
-  prefix: 'jp'
-}
+  param: "jsonpCallback",
+  prefix: "jp",
+};
 
 // request 成功取回状态码
-export const ERR_OK = 0
+export const ERR_OK = 0;
