@@ -1,81 +1,81 @@
-import axios from 'axios'
-import { commonParams } from './config'
+import axios from "axios";
+import { commonParams } from "./config";
 
-const debug = process.env.NODE_ENV !== 'production'
+const debug = process.env.NODE_ENV !== "production";
 
 // 获取推荐页面轮播图数据
 export function getSliderList() {
   const url = debug
-    ? '/api/getSliderList'
-    : 'http://127.0.0.1:9095/api/getSliderList'
+    ? "/api/getSliderList"
+    : "http://127.0.0.1:9095/api/getSliderList";
 
   const data = Object.assign({}, commonParams, {
-    platform: 'yqq.json',
+    platform: "yqq.json",
     hostUin: 0,
     needNewCode: 0,
-    inCharset: 'utf8',
-    format: 'json',
-    '-': 'recom' + (Math.random() + '').replace('0.', ''),
+    inCharset: "utf8",
+    format: "json",
+    "-": "recom" + (Math.random() + "").replace("0.", ""),
     data: {
       comm: { ct: 24 },
       category: {
-        method: 'get_hot_category',
-        param: { qq: '' },
-        module: 'music.web_category_svr'
+        method: "get_hot_category",
+        param: { qq: "" },
+        module: "music.web_category_svr",
       },
       recomPlaylist: {
-        method: 'get_hot_recommend',
+        method: "get_hot_recommend",
         param: { async: 1, cmd: 2 },
-        module: 'playlist.HotRecommendServer'
+        module: "playlist.HotRecommendServer",
       },
       playlist: {
-        method: 'get_playlist_by_category',
+        method: "get_playlist_by_category",
         param: { id: 8, curPage: 1, size: 40, order: 5, titleid: 8 },
-        module: 'playlist.PlayListPlazaServer'
+        module: "playlist.PlayListPlazaServer",
       },
       new_song: {
-        module: 'newsong.NewSongServer',
-        method: 'get_new_song_info',
-        param: { type: 5 }
+        module: "newsong.NewSongServer",
+        method: "get_new_song_info",
+        param: { type: 5 },
       },
       new_album: {
-        module: 'newalbum.NewAlbumServer',
-        method: 'get_new_album_info',
-        param: { area: 1, sin: 0, num: 10 }
+        module: "newalbum.NewAlbumServer",
+        method: "get_new_album_info",
+        param: { area: 1, sin: 0, num: 10 },
       },
       new_album_tag: {
-        module: 'newalbum.NewAlbumServer',
-        method: 'get_new_album_area',
-        param: {}
+        module: "newalbum.NewAlbumServer",
+        method: "get_new_album_area",
+        param: {},
       },
       toplist: {
-        module: 'musicToplist.ToplistInfoServer',
-        method: 'GetAll',
-        param: {}
+        module: "musicToplist.ToplistInfoServer",
+        method: "GetAll",
+        param: {},
       },
       focus: {
-        module: 'QQMusic.MusichallServer',
-        method: 'GetFocus',
-        param: {}
-      }
-    }
-  })
+        module: "QQMusic.MusichallServer",
+        method: "GetFocus",
+        param: {},
+      },
+    },
+  });
 
   return axios
     .get(url, {
-      params: data
+      params: data,
     })
-    .then(res => res.data)
+    .then((res) => res.data);
 }
 
 // 获取推荐页面歌单数据
 export function getDiscList() {
   const url = debug
-    ? '/api/getDiscList'
-    : 'http://127.0.0.1:9095/api/getDiscList'
+    ? "/api/getDiscList"
+    : "http://127.0.0.1:9095/api/getDiscList";
 
   const data = Object.assign({}, commonParams, {
-    platform: 'yqq',
+    platform: "yqq",
     hostUin: 0,
     needNewCode: 0,
     sin: 0,
@@ -83,33 +83,33 @@ export function getDiscList() {
     sortId: 5,
     categoryId: 10000000,
     rnd: Math.random(),
-    format: 'json'
-  })
+    format: "json",
+  });
 
   return axios
     .get(url, {
-      params: data
+      params: data,
     })
-    .then(res => res.data)
+    .then((res) => res.data);
 }
 
 // 获取歌单歌曲数据
 export function getSongList(disstid) {
-  const url = debug ? '/api/getCdInfo/' : 'http://127.0.0.1:9095/api/getCdInfo'
+  const url = debug ? "/api/getCdInfo/" : "http://127.0.0.1:9095/api/getCdInfo";
   const data = Object.assign({}, commonParams, {
     disstid,
     type: 1,
     json: 1,
     utf8: 1,
     onlysong: 0,
-    platform: 'yqq',
+    platform: "yqq",
     hostUin: 0,
-    needNewCode: 0
-  })
+    needNewCode: 0,
+  });
 
   return axios
     .get(url, {
-      params: data
+      params: data,
     })
-    .then(res => res.data)
+    .then((res) => res.data);
 }
