@@ -1,20 +1,20 @@
-import axios from 'axios'
-import jsonp from '../utils/jsonp'
-import { commonParams, options } from './config'
+import axios from "axios";
+import jsonp from "../utils/jsonp";
+import { commonParams, options } from "./config";
 
-const debug = process.env.NODE_ENV !== 'production'
+const debug = process.env.NODE_ENV !== "production";
 
 // 获取搜索关键词
 export function getHotKey() {
-  const url = 'https://c.y.qq.com/splcloud/fcgi-bin/gethotkey.fcg'
+  const url = "https://c.y.qq.com/splcloud/fcgi-bin/gethotkey.fcg";
 
   const data = Object.assign({}, commonParams, {
     uin: 0,
-    platform: 'h5',
-    needNewCode: 1
-  })
+    platform: "h5",
+    needNewCode: 1,
+  });
 
-  return jsonp(url, data, options)
+  return jsonp(url, data, options);
 }
 
 /**
@@ -25,7 +25,7 @@ export function getHotKey() {
  * @param {String} perpage 每页数目
  */
 export function search(query, page, zhida, perpage) {
-  const url = debug ? '/api/search' : 'http://127.0.0.1:9095/api/search'
+  const url = debug ? "/api/search" : "http://127.0.0.1:9095/api/search";
 
   const data = Object.assign({}, commonParams, {
     w: query,
@@ -36,19 +36,19 @@ export function search(query, page, zhida, perpage) {
     zhidaqu: 1,
     t: 0,
     flag: 1,
-    ie: 'utf-8',
+    ie: "utf-8",
     sem: 1,
     aggr: 0,
-    remoteplace: 'txt.mqq.all',
+    remoteplace: "txt.mqq.all",
     uin: 0,
-    platform: 'h5',
-    format: 'json',
-    needNewCode: 1
-  })
+    platform: "h5",
+    format: "json",
+    needNewCode: 1,
+  });
 
   return axios
     .get(url, {
-      params: data
+      params: data,
     })
-    .then(res => res.data)
+    .then((res) => res.data);
 }
