@@ -1,38 +1,38 @@
-import jsonp from '../utils/jsonp'
-import { commonParams, options } from './config'
+import jsonp from "../utils/jsonp";
+import { commonParams, options } from "./config";
 
 // 获取歌手列表
 export function getSingerList() {
-  const url = 'https://c.y.qq.com/v8/fcg-bin/v8.fcg'
+  const url = "https://c.y.qq.com/v8/fcg-bin/v8.fcg";
 
   const data = Object.assign({}, commonParams, {
-    channel: 'singer',
-    page: 'list',
-    key: 'all_all_all',
+    channel: "singer",
+    page: "list",
+    key: "all_all_all",
     pagesize: 100,
     pagenum: 1,
     hostUin: 0,
     needNewCode: 0,
-    platform: 'yqq'
-  })
+    platform: "yqq",
+  });
 
-  return jsonp(url, data, options)
+  return jsonp(url, data, options);
 }
 
 // 获取歌手详情
 export function getSingerDetail(singerId) {
-  const url = 'https://c.y.qq.com/v8/fcg-bin/fcg_v8_singer_track_cp.fcg'
+  const url = "https://c.y.qq.com/v8/fcg-bin/fcg_v8_singer_track_cp.fcg";
 
   const data = Object.assign({}, commonParams, {
     hostUin: 0,
     needNewCode: 0,
-    platform: 'yqq',
-    order: 'listen',
+    platform: "yqq",
+    order: "listen",
     begin: 0,
     num: 100,
     songstatus: 1,
-    singermid: singerId
-  })
+    singermid: singerId,
+  });
 
-  return jsonp(url, data, options)
+  return jsonp(url, data, options);
 }
