@@ -1,23 +1,23 @@
-export const SET_SINGER = 'SET_SINGER'
+export const SET_SINGER = "SET_SINGER";
 
-export const SET_PLAYING_STATE = 'SET_PLAYING_STATE'
+export const SET_PLAYING_STATE = "SET_PLAYING_STATE";
 
-export const SET_FULL_SCREEN = 'SET_FULL_SCREEN'
+export const SET_FULL_SCREEN = "SET_FULL_SCREEN";
 
-export const SET_PLAYLIST = 'SET_PLAYLIST'
+export const SET_PLAYLIST = "SET_PLAYLIST";
 
-export const SET_SEQUENCE_LIST = 'SET_SEQUENCE_LIST'
+export const SET_SEQUENCE_LIST = "SET_SEQUENCE_LIST";
 
-export const SET_PLAY_MODE = 'SET_PLAY_MODE'
+export const SET_PLAY_MODE = "SET_PLAY_MODE";
 
-export const SET_CURRENT_INDEX = 'SET_CURRENT_INDEX'
+export const SET_CURRENT_INDEX = "SET_CURRENT_INDEX";
 
-export const SET_DISC = 'SET_DISC'
+export const SET_DISC = "SET_DISC";
 
-export const SET_TOP_LIST = 'SET_TOP_LIST'
+export const SET_TOP_LIST = "SET_TOP_LIST";
 
-export const SET_SEARCH_HISTORY = 'SET_SEARCH_HISTORY'
+export const SET_SEARCH_HISTORY = "SET_SEARCH_HISTORY";
 
-export const SET_PLAY_HISTORY = 'SET_PLAY_HISTORY'
+export const SET_PLAY_HISTORY = "SET_PLAY_HISTORY";
 
-export const SET_FAVORITE_LIST = 'SET_FAVORITE_LIST'
+export const SET_FAVORITE_LIST = "SET_FAVORITE_LIST";
