@@ -2,5 +2,5 @@
 export const playMode = {
   sequence: 0, // 顺序播放
   loop: 1, // 单曲循环
-  random: 2 // 随机播放
-}
+  random: 2, // 随机播放
+};
