@@ -1,12 +1,12 @@
-let uid = ''
+let uid = "";
 
 export function getUid() {
   if (uid) {
-    return uid
+    return uid;
   }
   if (!uid) {
-    const t = new Date().getUTCMilliseconds()
-    uid = '' + ((Math.round(2147483647 * Math.random()) * t) % 1e10)
+    const t = new Date().getUTCMilliseconds();
+    uid = "" + ((Math.round(2147483647 * Math.random()) * t) % 1e10);
   }
-  return uid
+  return uid;
 }
